requirements.txt
Django==1.11.2
djangorestframework==3.6.3
pytz==2017.2
virtualenv==15.1.0
